Added HP button
Aaron's idea was to show enemy hp whenever the mouse was over that
enemy. After trying to do this, I realized it wasn't that simple since
the input processor switches between the stage(UI) and the map. Also, we
would probably have to check every enemy constantly to see if the mouse
position was at one of them every time the mouse was moved, which would
get expensive when VERY quickly. My solution was to add a button that
can be held down to show the health of every enemy on the map. If anyone
is unhappy with it, they might want to look into Input Multiplexers and
think of an efficient way to check mouse vs. enemy location.